// frontend/plugins/welcome/src/api/models/EntManager.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntManagerEdges,
    EntManagerEdgesFromJSON,
    EntManagerEdgesFromJSONTyped,
    EntManagerEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntManager
 */
export interface EntManager {
    /**
     * 
     * @type {EntManagerEdges}
     * @memberof EntManager
     */
    edges?: EntManagerEdges;
    /**
     * Email holds the value of the "email" field.
     * @type {string}
     * @memberof EntManager
     */
    email?: string;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntManager
     */
    id?: number;
    /**
     * Name holds the value of the "name" field.
     * @type {string}
     * @memberof EntManager
     */
    name?: string;
    /**
     * Password holds the value of the "password" field.
     * @type {string}
     * @memberof EntManager
     */
    password?: string;
}

export function EntManagerFromJSON(json: any): EntManager {
    return EntManagerFromJSONTyped(json, false);
}

export function EntManagerFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntManager {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'edges': !exists(json, 'edges') ? undefined : EntManagerEdgesFromJSON(json['edges']),
        'email': !exists(json, 'email') ? undefined : json['email'],
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'password': !exists(json, 'password') ? undefined : json['password'],
    };
}

export function EntManagerToJSON(value?: EntManager | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'edges': EntManagerEdgesToJSON(value.edges),
        'email': value.email,
        'id': value.id,
        'name': value.name,
        'password': value.password,
    };
}